Add null check in DbInitializer and update DbSet properties
Added a check to ensure that the context in the Initialize method of DbInitializer class is not null, throwing an exception if it is. Also changed the setters of DbSet properties in ApplicationDbContext class to be init-only which ensures they are set only at object creation time and cannot be modified afterwards.