Refactor UserTasks to separate logic and view
UserTasks component has been refactored to separate the business logic from the view. The existing logic defined in the UserTasks.razor file was moved to a newly created UserTasks.razor.cs file. The main benefit is clear separation and maintainability of the code in both the backend and frontend perspective.