Add BlazorMaybePocTests project and setup unit tests
This commit includes the creation of the BlazorMaybePocTests project and sets up its dependencies for NUnit, Moq, and others. Additionally, a simple UserTasks test class has been added as an example. Changes were also made to the UserTasks component in the Client project to make certain methods virtual for unit testing.