Add benchmark project and update temperature conversion service
A new 'benchmark' project has been added to the solution to measure the performance of the temperature conversion service. Some improvements have been made in the temperature conversion service, including a new method for the 'Maybe' implementation. This update will help in testing the efficiency and speed of the Celsius to Fahrenheit conversion.