Refactor TemperatureConversionService to use monads
The commit introduces a significant refactor of the TemperatureConversionService implementation. In addition to adding the "DotNext" package, relevant code changes are made to replace the previous 'Maybe' implementation with monads. The corresponding file and class names are also updated to reflect the use of monads. The benchmark tests were expanded to cover more cases including heap and stack allocation with delegates and lambdas.